style: fix RuboCop offenses — extract Structural spec, auto-correct alignment
- Run rubocop -A on all mirror specs (36 auto-corrections)
- Extract Structural handler specs into separate file (fix MultipleDescribes)
- Extract complex_doc into let (fix ExampleLength in roundtrip_spec)
- Add remaining_handlers_spec to rubocop_todo MultipleDescribes exclusion
