+++ spec/metanorma/mirror/handlers/table_handler_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Metanorma::Mirror::Handlers::Table do
   let(:id_strategy) { Metanorma::Mirror::IdStrategy::Preserve.new }
   let(:context) do
     Metanorma::Mirror::MetanormaToMirror.new(registry: registry,
-                                              id_strategy: id_strategy)
+                                             id_strategy: id_strategy)
   end
 
   def parse_table(xml)
